+++ packages/react-dom/types/Link.d.ts
@@ -1,16 +1,19 @@
 import React from "react";
 import { RouteLocation } from "@curi/types";
-import { NavigatingChildren } from "@curi/react-universal";
-export interface BaseLinkProps extends RouteLocation {
-    onNav?: (e: React.MouseEvent<HTMLElement>) => void;
+import { NavigatingChildren, NavigationHookProps } from "@curi/react-universal";
+interface BaseLinkProps extends RouteLocation {
     anchor?: React.ReactType;
-    forward?: React.AnchorHTMLAttributes<HTMLAnchorElement>;
 }
-export interface LinkProps extends BaseLinkProps {
+export declare type LinkProps = BaseLinkProps & NavigationHookProps<React.MouseEvent<HTMLElement>> & {
     children: React.ReactNode;
-}
-export interface AsyncLinkProps extends BaseLinkProps {
+};
+export declare type AsyncLinkProps = BaseLinkProps & NavigationHookProps<React.MouseEvent<HTMLElement>> & {
     children: NavigatingChildren;
-}
-export declare const Link: React.ForwardRefExoticComponent<LinkProps & React.RefAttributes<any>>;
-export declare const AsyncLink: React.ForwardRefExoticComponent<AsyncLinkProps & React.RefAttributes<any>>;
+};
+export declare const Link: React.ForwardRefExoticComponent<BaseLinkProps & NavigationHookProps<React.MouseEvent<HTMLElement, MouseEvent>> & {
+    children: React.ReactNode;
+} & React.RefAttributes<any>>;
+export declare const AsyncLink: React.ForwardRefExoticComponent<BaseLinkProps & NavigationHookProps<React.MouseEvent<HTMLElement, MouseEvent>> & {
+    children: NavigatingChildren;
+} & React.RefAttributes<any>>;
+export {};