+++ packages/react-universal/types/hooks/useNavigationHandler.d.ts
@@ -1,24 +1,20 @@
 import React from "react";
 import { ReactNode } from "react";
-import { RouteLocation } from "@curi/types";
 import { NavType } from "@hickory/root";
-export declare type NavigatingChildren = (navigating: boolean) => ReactNode;
-export interface NavigationHookProps<T> extends BaseNavigationHookProps<T> {
-    children: React.ReactNode;
-}
-export interface StatefulNavigationHookProps<T> extends BaseNavigationHookProps<T> {
-    children: NavigatingChildren;
-}
-export interface BaseNavigationHookProps<T> extends RouteLocation {
+export interface NavigationHookProps<T> {
     onNav?: (e: T) => void;
-    forward?: object;
     method?: NavType;
+    forward?: object;
+    state?: any;
 }
+export declare type NavigatingChildren = (navigating: boolean) => ReactNode;
 export declare type CanNavigate<T> = (e: T, target?: string) => boolean;
-export declare function useNavigationHandler<T extends React.BaseSyntheticEvent>(props: NavigationHookProps<T>, canNavigate?: CanNavigate<T>, target?: string): {
+export declare function useNavigationHandler<T extends React.BaseSyntheticEvent>(url: string, props: NavigationHookProps<T>, canNavigate?: CanNavigate<T>): {
+    url: string;
     eventHandler: (event: T) => void;
 };
-export declare function useStatefulNavigationHandler<T extends React.BaseSyntheticEvent>(props: StatefulNavigationHookProps<T>, canNavigate?: CanNavigate<T>, target?: string): {
+export declare function useStatefulNavigationHandler<T extends React.BaseSyntheticEvent>(url: string, props: NavigationHookProps<T>, canNavigate?: CanNavigate<T>): {
+    url: string;
     eventHandler: (event: T) => void;
     navigating: boolean;
 };