+++ packages/vue/src/Link.ts
@@ -2,7 +2,6 @@ import Vue from "vue";
 import { canNavigate } from "./utils/canNavigate";
 
 import { CreateElement, ComponentOptions } from "vue";
-import { SessionLocation } from "@hickory/root";
 
 export interface LinkComponent extends Vue {
   name: string;
@@ -10,8 +9,7 @@ export interface LinkComponent extends Vue {
   hash?: string;
   query?: any;
   state?: any;
-  location: SessionLocation;
-  href: string;
+  url: string;
   click(e: MouseEvent): void;
   forward?: object;
 }
@@ -22,19 +20,13 @@ const Link: ComponentOptions<LinkComponent> = {
   props: ["name", "params", "hash", "query", "state", "click", "forward"],
 
   computed: {
-    location: function() {
-      const pathname = this.name
-        ? this.$router.route.pathname(this.name, this.params)
-        : "";
-      return {
+    url: function() {
+      return this.$router.url({
+        name: this.name,
+        params: this.params,
         hash: this.hash,
-        query: this.query,
-        state: this.state,
-        pathname
-      };
-    },
-    href: function() {
-      return this.$router.history.href(this.location);
+        query: this.query
+      });
     }
   },
 
@@ -43,13 +35,11 @@ const Link: ComponentOptions<LinkComponent> = {
       if (this.click) {
         this.click(event);
       }
-      if (canNavigate(event)) {
+      // @ts-ignore
+      if (canNavigate(event, this.forward && this.forward.target)) {
         event.preventDefault();
         this.$router.navigate({
-          name: this.name,
-          params: this.params,
-          hash: this.hash,
-          query: this.query,
+          url: this.url,
           state: this.state
         });
       }
@@ -60,7 +50,7 @@ const Link: ComponentOptions<LinkComponent> = {
     return h(
       "a",
       {
-        attrs: { href: this.href, ...this.forward },
+        attrs: { href: this.url, ...this.forward },
         on: { click: this.clickHandler }
       },
       this.$slots.default