+++ packages/vue/src/utils/canNavigate.ts
@@ -1,7 +1,8 @@
-export function canNavigate(event: MouseEvent) {
+export function canNavigate(event: MouseEvent, target?: string) {
   return (
     !event.defaultPrevented &&
     (event.button !== undefined && event.button === 0) &&
-    !(event.metaKey || event.altKey || event.ctrlKey || event.shiftKey)
+    !(event.metaKey || event.altKey || event.ctrlKey || event.shiftKey) &&
+    (!target || target === "_self")
   );
 }