+++ packages/vue/types/AsyncLink.d.ts
@@ -1,14 +1,12 @@
 import Vue from "vue";
 import { ComponentOptions } from "vue";
-import { SessionLocation } from "@hickory/root";
 export interface AsyncLinkComponent extends Vue {
     name: string;
     params?: object;
     hash?: string;
     query?: any;
     state?: any;
-    location: SessionLocation;
-    href: string;
+    url: string;
     click(e: MouseEvent): void;
     navigating: boolean;
     forward?: object;