+++ packages/svelte/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+* Use `router.url`.
+
 ## 2.0.0-alpha.0
 
 * Bump to `v2` alpha since it uses `@curi/router` v2.